+++ R/estimate_bootstrap.R
@@ -71,10 +71,14 @@ bootstrap_model <- function(seminr_model, nboot = 500, cores = NULL, seed = NULL
         # Initialize the cluster
         suppressWarnings(ifelse(is.null(cores), cl <- parallel::makeCluster(parallel::detectCores(), setup_strategy = "sequential"), cl <- parallel::makeCluster(cores, setup_strategy = "sequential")))
 
-        # Initialize the Estimates Matrix
-        bootstrapMatrix <- rbind(seminr_model$path_coef, seminr_model$outer_loadings, seminr_model$outer_weights, HTMT(seminr_model), total_effects(seminr_model$path_coef))
-        cols <- ncol(bootstrapMatrix)
-        rows <- nrow(bootstrapMatrix)
+        # # Initialize the Estimates Matrix
+        # bootstrapMatrix <- as.matrix(c(c(seminr_model$path_coef),
+        #                              c(seminr_model$outer_loadings),
+        #                              c(seminr_model$outer_weights),
+        #                              c(seminr:::HTMT(seminr_model)),
+        #                              c(seminr:::total_effects(seminr_model$path_coef))))
+        path_cols <- ncol(seminr_model$path_coef)
+        path_rows <- nrow(seminr_model$path_coef)
 
         # Function to generate random samples with replacement
         getRandomIndex <- function(d) {return(sample.int(nrow(d), replace = TRUE))}
@@ -94,22 +98,37 @@ bootstrap_model <- function(seminr_model, nboot = 500, cores = NULL, seed = NULL
                                inner_weights)
           boot_htmt <- HTMT(boot_model)
           boot_total <- total_effects(boot_model$path_coef)
-          return(rbind(boot_model$path_coef, boot_model$outer_loadings, boot_model$outer_weights, boot_htmt, boot_total))
+          return(as.matrix(c(c(boot_model$path_coef),
+                             c(boot_model$outer_loadings),
+                             c(boot_model$outer_weights),
+                             c(boot_htmt),
+                             c(boot_total))))
         }
 
         # Bootstrap the estimates
         utils::capture.output(bootmatrix <- parallel::parSapply(cl, 1:nboot, getEstimateResults, d))
 
         # Add the columns for bootstrap mean and standard error
-        bootstrapMatrix <- cbind(bootstrapMatrix, matrix(apply(bootmatrix,1,mean), nrow = rows, ncol = cols))
-        bootstrapMatrix <- cbind(bootstrapMatrix, matrix(apply(bootmatrix,1,stats::sd), nrow = rows, ncol = cols))
+        means <- apply(bootmatrix,1,mean)
+        sds <- apply(bootmatrix,1,stats::sd)
+        start <- 1
+        end <- (path_cols*path_rows)
+
+        # Create the array of bootstrap paths
+        boot_paths <- array(bootmatrix[start:end,1:nboot], dim = c(path_rows, path_cols, nboot), dimnames = list(rownames(seminr_model$path_coef), colnames(seminr_model$path_coef),1:nboot))
+
+        # Create the summary matrices of means and sds
+        paths_means <- matrix(means[start:end], nrow = path_rows, ncol = path_cols)
+        paths_sds <- matrix(sds[start:end], nrow = path_rows, ncol = path_cols)
 
         # Subset paths matrix (take care to not be stranded with a single column/row vector)
-        paths_descriptives <- matrix(bootstrapMatrix[1:(cols-1), c(1:(3*cols))], nrow=(cols-1), ncol=(3*cols))
-        dimnames(paths_descriptives) <- list(
-          rownames(bootstrapMatrix)[1:(cols-1)],
-          colnames(bootstrapMatrix)[1:(3*cols)]
-        )
+        paths_descriptives <- cbind(seminr_model$path_coef, paths_means, paths_sds)
+
+        # Apply dimnames to matrix
+        # dimnames(paths_descriptives) <- list(
+          # rownames(seminr_model$path_coef),
+          # colnames(bootstrapMatrix)[1:(3*cols)]
+        # )
 
         # Clean the empty paths (take care to not be stranded with a single column/row vector)
         filled_cols <- apply(paths_descriptives != 0, 2, any, na.rm=TRUE)
@@ -132,18 +151,37 @@ bootstrap_model <- function(seminr_model, nboot = 500, cores = NULL, seed = NULL
           }
         }
 
+        # # construct vector of rownames
+        # independant <- unique(structural_model[,"source"])
         # Assign column names
         colnames(paths_descriptives) <- col_names
 
         # collect loadings matrix
-        loadings_descriptives <- bootstrapMatrix[(cols+1):((cols)+nrow(seminr_model$outer_loadings)), c(1:(3*cols))]
+        mm_cols <- ncol(seminr_model$outer_loadings)
+        mm_rows <- nrow(seminr_model$outer_loadings)
+        start <- end+1
+        end <- start+(mm_cols*mm_rows)-1
 
+        # create the array of bootstrapped loadings
+        boot_loadings <- array(bootmatrix[start:end,1:nboot], dim = c(mm_rows, mm_cols, nboot), dimnames = list(rownames(seminr_model$outer_loadings), colnames(seminr_model$outer_loadings),1:nboot))
+
+        # Create the summary matrices of means and sds
+        loadings_means <- matrix(means[start:end], nrow = mm_rows, ncol = mm_cols)
+        loadings_sds <- matrix(sds[start:end], nrow = mm_rows, ncol = mm_cols)
+
+        # Subset paths matrix (take care to not be stranded with a single column/row vector)
+        loadings_descriptives <- cbind(seminr_model$outer_loadings, loadings_means, loadings_sds)
+
+        # loadings_descriptives <- bootstrapMatrix[(cols-1):((cols-2)+nrow(seminr_model$outer_loadings)), c(1:(3*cols))]
+
+        # Apply rownames to matrix
+        # rownames(loadings_descriptives) <- rownames(seminr_model$outer_loadings)
 
         # Construct the vector of column names 2
         col_names2 <- c()
         # Clean the column names
         for (parameter in c("PLS Est.", "Boot Mean", "Boot SD")) {
-          for(i in seminr_model$constructs) {
+          for(i in colnames(seminr_model$outer_loadings)) {
             col_names2 <- c(col_names2, paste(i, parameter, sep = " "))
           }
         }
@@ -152,27 +190,79 @@ bootstrap_model <- function(seminr_model, nboot = 500, cores = NULL, seed = NULL
         colnames(loadings_descriptives) <- col_names2
 
         # collect weights matrix
-        weights_descriptives <- bootstrapMatrix[((cols+1)+nrow(seminr_model$outer_loadings)):((cols)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))]
+        start <- end+1
+        end <- start+(mm_cols*mm_rows)-1
+
+        # create the array of bootstrapped weights
+        boot_weights <- array(bootmatrix[start:end,1:nboot], dim = c(mm_rows, mm_cols, nboot), dimnames = list(rownames(seminr_model$outer_loadings), colnames(seminr_model$outer_loadings),1:nboot))
+
+        # Create the summary matrices of means and sds
+        weights_means <- matrix(means[start:end], nrow = mm_rows, ncol = mm_cols)
+        weights_sds <- matrix(sds[start:end], nrow = mm_rows, ncol = mm_cols)
+
+        # Subset paths matrix (take care to not be stranded with a single column/row vector)
+        weights_descriptives <- cbind(seminr_model$outer_weights, weights_means, weights_sds)
+        # weights_descriptives <- bootstrapMatrix[((cols-1)+nrow(seminr_model$outer_loadings)):((cols-2)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))]
+
+        # Apply rownames to matrix
+        # rownames(weights_descriptives) <- rownames(seminr_model$outer_loadings)
 
         # Assign column names to weights
         colnames(weights_descriptives) <- col_names2
 
         # Collect HTMT matrix
-        HTMT_descriptives <- bootstrapMatrix[((cols+1)+(2*nrow(seminr_model$outer_loadings))):((cols+cols)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))]
+        HTMT_matrix <- seminr:::HTMT(seminr_model)
+        start <- end+1
+        end <- start+(path_cols*path_rows)-1
+
+        # Collect the array of bootstrapped HTMT
+        boot_HTMT <- array(bootmatrix[start:end,1:nboot], dim = c(path_rows, path_cols, nboot), dimnames = list(rownames(HTMT_matrix), colnames(HTMT_matrix),1:nboot))
+
+        # Collect the matrices of means and sds
+        htmt_means <- matrix(means[start:end], nrow = path_rows, ncol = path_cols)
+        htmt_sds <- matrix(sds[start:end], nrow = path_rows, ncol = path_cols)
+
+
+        # Subset paths matrix (take care to not be stranded with a single column/row vector)
+        HTMT_descriptives <- cbind(HTMT_matrix, htmt_means, htmt_sds)
+        # HTMT_descriptives <- bootstrapMatrix[((cols-1)+(2*nrow(seminr_model$outer_loadings))):((cols+cols-4)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))]
 
         # Clean the empty paths
         #boot_HTMT <- boot_HTMT[, colSums(boot_HTMT != 0, na.rm = TRUE) > 0]
         #boot_HTMT <- boot_HTMT[rowSums(boot_HTMT != 0, na.rm = TRUE) > 0,]
 
+        # Construct the vector of column names 3 for HTMT
+        col_names3 <- c()
+        # Clean the column names
+        for (parameter in c("PLS Est.", "Boot Mean", "Boot SD")) {
+          for(i in colnames(HTMT_matrix)) {
+            col_names3 <- c(col_names3, paste(i, parameter, sep = " "))
+          }
+        }
         # Get boot_HTMT column names
-        colnames(HTMT_descriptives) <- col_names2
+        colnames(HTMT_descriptives) <- col_names3
 
         # Subset total paths matrix (take care to not be stranded with a single column/row vector)
-        total_paths_descriptives <- matrix(bootstrapMatrix[((2*cols+1)+2*nrow(seminr_model$outer_loadings)):((3*cols)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))], nrow=(cols), ncol=(3*cols))
-        dimnames(total_paths_descriptives) <- list(
-          rownames(bootstrapMatrix)[1:(cols)],
-          colnames(bootstrapMatrix)[1:(3*cols)]
-        )
+        total_matrix <- seminr:::total_effects(seminr_model$path_coef)
+        start <- end+1
+        end <- start+(path_cols*path_rows)-1
+
+        # Collect the array of bootstrapped total paths
+        boot_total_paths <- array(bootmatrix[start:end,1:nboot], dim = c(path_rows, path_cols, nboot), dimnames = list(rownames(total_matrix), colnames(total_matrix),1:nboot))
+
+        # Collect the matrices for means and sds
+        total_means <- matrix(means[start:end], nrow = path_rows, ncol = path_cols)
+        total_sds <- matrix(sds[start:end], nrow = path_rows, ncol = path_cols)
+
+
+        # Subset paths matrix (take care to not be stranded with a single column/row vector)
+        total_paths_descriptives <- cbind(total_matrix, total_means, total_sds)
+
+        # # total_paths_descriptives <- matrix(bootstrapMatrix[((2*cols+1)+2*nrow(seminr_model$outer_loadings)):((3*cols)+(2*nrow(seminr_model$outer_loadings))), c(1:(3*cols))], nrow=(cols), ncol=(3*cols))
+        # dimnames(total_paths_descriptives) <- list(
+        #   rownames(bootstrapMatrix)[1:(cols)],
+        #   colnames(bootstrapMatrix)[1:(3*cols)]
+        # )
 
         # Clean the empty paths (take care to not be stranded with a single column/row vector)
         filled_cols <- apply(total_paths_descriptives != 0, 2, any, na.rm=TRUE)
@@ -183,14 +273,16 @@ bootstrap_model <- function(seminr_model, nboot = 500, cores = NULL, seed = NULL
         colnames(total_paths_descriptives) <- col_names
 
         # Create an array of results in bootmatrix
-        bootarray <- array(bootmatrix, dim = c(nrow(bootstrapMatrix), length(seminr_model$constructs),nboot), dimnames = list(c(rownames(bootstrapMatrix)), c(seminr_model$constructs), c(1:nboot)))
+        # bootarray <- array(bootmatrix, dim = c(nrow(bootmatrix), length(seminr_model$constructs),nboot), dimnames = list(c(rownames(bootstrapMatrix)), c(seminr_model$constructs), c(1:nboot)))
 
         # Create arrays of bootstrapped path, loadings, weights, HTMT coefficients results
-        boot_paths <- bootarray[1:length(seminr_model$constructs), , 1:nboot]
-        boot_loadings <- bootarray[(length(seminr_model$constructs)+1):((length(seminr_model$constructs))+length(seminr_model$mmVariables)), , 1:nboot]
-        boot_weights <- bootarray[(length(seminr_model$constructs)+length(seminr_model$mmVariables)+1):((length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
-        boot_HTMT <- bootarray[((length(seminr_model$constructs))+2*length(seminr_model$mmVariables)+1):((2*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
-        boot_total_paths <- bootarray[((2*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)+1):((3*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
+        # boot_paths <- array(bootmatrix[1:(path_cols*path_rows),1:nboot], dim = c(path_rows, path_cols, nboot), dimnames = list(rownames(seminr_model$path_coef), colnames(seminr_model$path_coef),1:nboot))
+        # boot_paths <- bootarray[1:length(seminr_model$constructs), , 1:nboot]
+        # boot_loadings <- bootarray[(length(seminr_model$constructs)+1):((length(seminr_model$constructs))+length(seminr_model$mmVariables)), , 1:nboot]
+        # boot_loadings <- array(bootmatrix[1:(path_cols*path_rows),1:nboot], dim = c(path_rows, path_cols, nboot), dimnames = list(rownames(seminr_model$path_coef), colnames(seminr_model$path_coef),1:nboot))
+        # boot_weights <- bootarray[(length(seminr_model$constructs)+length(seminr_model$mmVariables)+1):((length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
+        # boot_HTMT <- bootarray[((length(seminr_model$constructs))+2*length(seminr_model$mmVariables)+1):((2*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
+        # boot_total_paths <- bootarray[((2*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)+1):((3*length(seminr_model$constructs))+2*length(seminr_model$mmVariables)), , 1:nboot]
 
         # Change the class of the descriptives objects
         class(paths_descriptives) <- append(class(paths_descriptives), "table_output")